updt routes, views, render comments

+++ public/js/edit.js
@@ -1,32 +1,50 @@
-const postId = document.querySelector('input[name="post-id"]').value;
+// const postId = document.querySelector("#postnum").value;
 
-const editFormHandler = async function(event) {
+const editFormHandler = async (event) => {
   event.preventDefault();
 
-  const title = document.querySelector('#title').value;
-  const body = document.querySelector('#content').value;
+  const id = event.target.getAttribute("data-id");
+  const newTitle = document.querySelector(`#title${id}`).value;
+  const newDescription = document.querySelector(`#content${id}`).value;
 
-  await fetch(`/api/post/${postId}`, {
-    method: 'PUT',
+  await fetch(`/api/posts/edit/${id}`, {
+    method: "PUT",
     body: JSON.stringify({
-      title,
-      body
+      title: newTitle,
+      description: newDescription,
     }),
     headers: {
-      'Content-Type': 'application/json'
+      "Content-Type": "application/json",
+    },
+  });
+
+  alert("Successfully updated post");
+  document.location.reload();
+};
+
+const delButtonHandler = async (event) => {
+  if (event.target.hasAttribute("data-id")) {
+    const id = event.target.getAttribute("data-id");
+
+    const response = await fetch(`/api/posts/${id}`, {
+      method: "DELETE",
+    });
+
+    if (response.ok) {
+      alert("Successfully deleted post");
+      document.location.reload();
+    } else {
+      alert("Failed to delete post");
     }
-  });
-
-  document.location.replace('/dashboard');
+  }
 };
 
-const deleteClickHandler = async function() {
-  await fetch(`/api/post/${postId}`, {
-    method: 'DELETE'
-  });
+const deleteBTN = document.querySelectorAll(".delete-btn");
+for (let i = 0; i < deleteBTN.length; i++) {
+  deleteBTN[i].addEventListener("click", delButtonHandler);
+}
 
-  document.location.replace('/dashboard');
-};
-
-document.querySelector('#update-btn').addEventListener('submit', editFormHandler);
-document.querySelector('#delete-btn').addEventListener('click', deleteClickHandler);
+const updatebtn = document.querySelectorAll("#update-btn");
+for (let i = 0; i < updatebtn.length; i++) {
+  updatebtn[i].addEventListener("click", editFormHandler);
+}
